Fixed the file names of models, changed the definitions used for include guards to avoid collisions with the model.h file.
Also fixed release mode and removed any requirement for DirectX 12 libraries when used in release.

Fixed a few warning messages about unused variables too.

+++ Prio Engine/Graphics.cpp
@@ -11,7 +11,7 @@ CGraphics::CGraphics()
 	mpLight = nullptr;
 	mpDiffuseLightShader = nullptr;
 
-	mFieldOfView = D3DX_PI / 4;
+	mFieldOfView = static_cast<float>(D3DX_PI / 4);
 }
 
 CGraphics::~CGraphics()
